Bugfix for writing 64 bit integers as packet bytes

+++ Sadie.Shared/Networking/Packets/NetworkPacketWriter.cs
@@ -35,7 +35,7 @@ public class NetworkPacketWriter
 
     protected void WriteLong(long data)
     {
-        WriteBytes(BitConverter.GetBytes(data), true);
+        WriteBytes(BitConverter.GetBytes((int)data), true);
     }
 
     protected void WriteBool(bool boolean)
